first round of addressing comments & make ci pass
