+++ ddtrace/appsec/_handlers.py
@@ -8,7 +8,6 @@ from typing import Tuple
 from typing import Union
 
 from ddtrace._trace.span import Span
-from ddtrace._trace.trace_handlers import httpx_url_to_str
 from ddtrace.appsec._asm_request_context import _call_waf
 from ddtrace.appsec._asm_request_context import _call_waf_first
 from ddtrace.appsec._asm_request_context import _get_asm_context
@@ -25,6 +24,7 @@ from ddtrace.appsec._http_utils import normalize_headers
 from ddtrace.appsec._http_utils import parse_http_body
 from ddtrace.appsec._utils import Block_config
 from ddtrace.contrib import trace_utils
+from ddtrace.contrib.internal.httpx.utils import httpx_url_to_str
 from ddtrace.contrib.internal.trace_utils_base import _get_request_header_user_agent
 from ddtrace.contrib.internal.trace_utils_base import _set_url_tag
 from ddtrace.ext import http
