+++ ddtrace/internal/core/event_hub.py
@@ -90,7 +90,7 @@ def dispatch(event_id: str, args: Tuple[Any, ...] = ()) -> None:
                 raise
 
 
-def dispatch_event(event, *extra_args):
+def dispatch_event(event, *extra_args) -> None:
     event_name = getattr(event, "event_name", None)
     if event_name is None:
         raise ValueError("Event must have an 'event_name' attribute")
